Removed Maven dependency for installing the wrapper file.
Removed embedded wrapper jar file and removed step from pipeline which adds it to the final build.

Added version checks to prevent unnecessary updates.

Added more version information to the outputs when running setup or update.

Added an SCA Agent installation step to fix a bug where running a new version of the packager multiple times would cause file locks. This was due to the packager trying to install the agent on first run.

Changed install locations for the cli and wrapper.