Wait for scans to complete and show results (#28)
Added functionality to wait for scans to complete and to show the results.

Updated dependencies.

Updated README and VHS scripts.

Fixed a bug with tasks that should run regardless of fatal errors not showing correctly on the reportcard.

Code refactors and cleanups.

Fixed minor bugs with error outputs not displaying on the reportcard.